Increase the timeout to avoid canceling the image building
At the last pipeline runs, the Image Builder
has taken longer for creating the OSTree Commit,
so it has expired the timeout.
It might be due network issue, but it's making
the pipeline break without aparent reason.
This hotfix tries to check if there's actually
an issue and let the PR create succefully piplines.